+++ Source/Classes/Completer/IconCompleter.cs
@@ -1,0 +1,38 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
+using System.Management.Automation;
+using System.Management.Automation.Abstractions;
+using System.Management.Automation.Language;
+
+namespace vsteam_lib
+{
+   public class IconCompleter : BaseCompleter
+   {
+      /// <summary>
+      /// This constructor is used when running in a PowerShell session. It cannot be
+      /// loaded in a unit test.
+      /// </summary>
+      [ExcludeFromCodeCoverage]
+      public IconCompleter() : base() { }
+
+      /// <summary>
+      /// This constructor is used during unit testings
+      /// </summary>
+      /// <param name="powerShell">fake instance of IPowerShell used for testing</param>
+      public IconCompleter(IPowerShell powerShell) : base(powerShell) { }
+
+      public override IEnumerable<CompletionResult> CompleteArgument(string commandName,
+                                                                     string parameterName,
+                                                                     string wordToComplete,
+                                                                     CommandAst commandAst,
+                                                                     IDictionary fakeBoundParameters)
+      {
+         var values = new List<CompletionResult>();
+
+         SelectValues(wordToComplete, IconCache.GetCurrent(false), values);
+
+         return values;
+      }
+   }
+}