+++ Source/Public/Set-VSTeamDefaultProject.ps1
@@ -49,19 +49,22 @@ function Set-VSTeamDefaultProject {
 
    process {
       if ($Force -or $pscmdlet.ShouldProcess($Project, "Set-VSTeamDefaultProject")) {
-         if (_isOnWindows) {
-            if (-not $Level) {
-               $Level = "Process"
-            }
+         #Set at the process level, for any OS
+         $env:TEAM_PROJECT = $Project
+         [vsteam_lib.Versions]::DefaultProject = $Project
 
-            # You always have to set at the process level or they will Not
-            # be seen in your current session.
-            $env:TEAM_PROJECT = $Project
-            [vsteam_lib.Versions]::DefaultProject = $Project
+         $env:TEAM_PROCESS = [vsteam_lib.Versions]::DefaultProcess =
+                _callapi -NoProject -area 'work' -resource 'processes' -QueryString @{'$expand'='projects'}  -version (_getAPIVersion Processes) |
+                   Select-Object -ExpandProperty Value  | 
+                     Where-Object {$_.psobject.properties['projects'] -and $_.projects.name -eq $ProjectName} |
+                        Select-Object -ExpandProperty Name 
 
-            [System.Environment]::SetEnvironmentVariable("TEAM_PROJECT", $Project, $Level)
+         if  ((_isOnWindows) -and $Level -and $level -ne "Process") {
+            [System.Environment]::SetEnvironmentVariable("TEAM_PROJECT", $Project,           $Level)
+            [System.Environment]::SetEnvironmentVariable("TEAM_PROCESS", $env:TEAM_PROCESS , $Level)
          }
-
+         
+         # Note: ProjectName Parameters should be given a default value of [vsteam_lib.Versions]::DefaultProject instead of globally forcing it this way.
          $Global:PSDefaultParameterValues["*-vsteam*:projectName"] = $Project
       }
    }